% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLGGM.R
\name{BLGGM}
\alias{BLGGM}
\title{Estimating Heterogeneous Gene Regulatory Networks from Zero-Inflated Single-Cell Expression Data}
\usage{
BLGGM(
  scRNA_data_matr,
  n_celltype,
  warm_cluster_label_init = FALSE,
  num_iterations = 10000,
  num_burnin = floor(num_iterations/2),
  collect_post_sample = FALSE,
  hyperparameters = c(0.2, 20, 0, 1, 3, -2, 0.2, 0.2, 0.01, 10, 0.02, 1, 1,
    2/(dim(scRNA_data_matr)[1] - 1)),
  hyperparameters_conc = rep(1, n_celltype),
  inclusion_thr = 0.5,
  print_label = FALSE,
  print_per_iteration = 1000,
  num_threads = 1
)
}
\arguments{
\item{scRNA_data_matr}{The scRNA-seq data matrix, where rows represent genes and columns represent cells. Matrix values need to be normalized single expression data.}

\item{n_celltype}{An integer, denoting the number of cell types.}

\item{warm_cluster_label_init}{The initialization of cluster labels is random or based on k-means. The default is FALSE, corresponding to the random initialization.}

\item{num_iterations}{The number of Gibbs sampler iterations. The default is 10000.}

\item{num_burnin}{The number of iterations in burn-in, after which the posterior samples are used to estimate the unknown parameters. The default is the first half of total iterations.}

\item{collect_post_sample}{Logical, collect the posterior samples or not. If users are only interested in the estimates, set it as FALSE to save the memory. If users would like to use posterior samples to construct credible intervals or for other uses, set it as TRUE. The default is FALSE.}

\item{hyperparameters}{A vector, which indicates 14 hyper-parameters in the priors or proposal distributions. The first two elements correspond to the step size and step number of leapfrog iteration for \eqn{\theta} respectively. The third and fourth elements are the mean and variance of the normal prior distribution for \eqn{\mu} respectively. The fifth and sixth elements are, respectively, the means of the normal priors for \eqn{\lambda_0} and \eqn{\lambda_1}. The seventh and eighth elements correspond to the variances of the normal priors for \eqn{\lambda_0} and \eqn{\lambda_1} respectively. The ninth and tenth elements are the step size and step number of leapfrog iteration for \eqn{\lambda} respectively. The eleventh and twelfth elements represent the variances of normal distributions in the spike-slab prior respectively. The thirteenth element is twice rate of exponential distributions in the spike-slab prior. The last one is the edge inclusion probability in the spike-slab prior. The default is 0.2, 20, 0, 1, 3, -2, 0.2, 0.2, 0.01, 10, 0.02, 1, 1, 2/(dim(scRNA_data_matr)[1] - 1)}

\item{hyperparameters_conc}{A vector is a vector of n_celltype dimensions, describing the parameters of Dirichlet prior for \eqn{\pi}. The default is a vector filled with one.}

\item{inclusion_thr}{A numeric between 0 and 1, denoting the threshold of edge inclusion. The edge with posterior probability of inclusion above the threshold is selected.}

\item{print_label}{Logical, whether or not to print summarized cell type labels after each iteration. The default is FALSE.}

\item{print_per_iteration}{An integer (denote it by v), print the iteration information per v iterations when print_label is FALSE. The default is 1000.}

\item{num_threads}{An integer, denoting the number of threads used in the parallel computing. The default is 1.}
}
\value{
BLGGM returns an R list including the following information.
\item{cell_labels}{A vector, indicating the estimated cell type labels for each cell.}
\item{cell_type_mean_expr}{A matrix of cell type expression profiles, in which rows are genes and columns correspond to cell types.}
\item{cell_type_precision_matr}{An array of cell-type-specific precision matrices, in which the third index corresponds to cell types.}
\item{cell_type_adj_matr}{An array of cell-type-specific gene adjacency matrix, in which the third index corresponds to cell types. 1 represents an edge, and 0 indicates no edge.}
\item{cell_type_edge_post_prob}{An array of cell-type-specific posterior probability matrices of edge inclusion, in which the third index corresponds to cell types.}
\item{prop}{A vector of estimated cell type proportions.}
\item{theta}{A matrix of \eqn{\theta}, in which rows are genes and columns correspond to cells.}
\item{lam0}{A vector, the estimated \eqn{\lambda_0} for each gene.}
\item{lam1}{A vector, the estimated \eqn{\lambda_1} for each gene.}
\item{cell_label_post}{Collected posterior samples of cell_labels when collect_post_sample is TRUE. If collect_post_sample is FALSE, this output does not exist.}
\item{cell_type_mean_expr_post}{Collected sampls of cell_type_mean_expr when collect_post_sample is TRUE. If collect_post_sample is FALSE, this output does not exist.}
\item{lam0_post}{Collected posterior samples of lam0 when collect_post_sample is TRUE. If collect_post_sample is FALSE, this output does not exist.}
\item{lam1_post}{Collected posterior samples of lam1 when collect_post_sample is TRUE. If collect_post_sample is FALSE, this output does not exist.}
\item{theta_post}{Collected posterior samples of \eqn{\theta} when collect_post_sample is TRUE. Rows are cells and columns are genes. If collect_post_sample is FALSE, this output does not exist.}
\item{cell_type_precision_matr_post}{Collected posterior samples of cell_type_precision_matr when collect_post_sample is TRUE. If collect_post_sample is FALSE, this output does not exist.}
\item{cell_type_edge_indicator_post}{Collected posterior samples of edge inclusion indicators for each cell type when collect_post_sample is TRUE. If collect_post_sample is FALSE, this output does not exist.}
}
\description{
The function BLGGM implements a Bayesian latent mixture GGM method to jointly estimate multiple gene regulatory networks accounting for the zero-inflation and unknown heterogeneity of single-cell expression data. It employs a hybrid Markov chain Monte Carlo algorithm to perform posterior inference.
}
\examples{
library(BLGGM)

#import example data
data(example_data)

#gene number
nrow(scRNA_data_matr)

#cell number
ncol(scRNA_data_matr)

#cell type number
n_celltype <- dim(precision_matr)[3]

#run BLGGM
set.seed(20200201)
t1 <- Sys.time()
Result <- BLGGM(scRNA_data_matr, n_celltype, num_iterations = 10000, num_threads = 10)
t2 <- Sys.time()

#time cost
print(t2 - t1)

#Compared with true cell type labels
cell_table <- table(Result$cell_labels, cell_label_truth)
cell_table

#The following shows the Frobenius norms of precision matrix within each cell type
sqrt(sum((Result$cell_type_precision_matr[,,which.max(cell_table[,1])] - precision_matr[,,1])^2))
sqrt(sum((Result$cell_type_precision_matr[,,which.max(cell_table[,2])] - precision_matr[,,2])^2))

}
\references{
Qiuyu Wu, and Xiangyu Luo. "Estimating Heterogeneous Gene Regulatory Networks from Zero-Inflated Single-Cell Expression Data."
}
